requirements.txt
python-docx
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv
langchain
langchain-google-genai
